Merge-Sort
We find the mid point of array 
Divide into left to mid and mid+1 to right
Create 2 new array and copy left and right to new arrays
Now compare in new arrays from their beginning and insert in original array from left to right in sorted fashion. 
If end of any of left and right array is not reached insert them into original array till it reaches end.