Count-Inversion
Similar to merge sort.
When we have left part and right part and left[i]>right[j]
since left and right are sorted, for all index in left >=i it will be greater than value of j so answer will be answer+sizeOf(left)-i and j++ as usual in merge-sort. This correspond to answer on this division.

when called mergeSort answer will sum of then two partition answers and answer from actual array we have that will half half sorted at end of recursion tree.